+++ CHANGELOG.md
@@ -92,6 +92,7 @@ in a future release.
     - `SqlServerDatabaseMail` was renamed to `SqlDatabaseMail`.
     - `SqlServerEndpoint` was renamed to `SqlEndpoint`.
     - `SqlServerEndpointPermission` was renamed to `SqlEndpointPermission`.
+    - `SqlServerLogin` was renamed to `SqlLogin`.
   - Changed all resource prefixes from `MSFT_` to `DSC_` ([issue #1496](https://github.com/dsccommunity/SqlServerDsc/issues/1496)).
     _Deprecated resource has not changed prefix._
   - All resources are now using the common module DscResource.Common.
@@ -184,7 +185,7 @@ in a future release.
   - BREAKING CHANGE: The parameter `ServerName` is now non-mandatory and
     defaults to `$env:COMPUTERNAME` ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).
   - Normalize parameter descriptive text for default values.
-- SqlServerLogin
+- SqlLogin
   - BREAKING CHANGE: The parameter `ServerName` is now non-mandatory and
     defaults to `$env:COMPUTERNAME` ([issue #319](https://github.com/dsccommunity/SqlServerDsc/issues/319)).
   - Normalize parameter descriptive text for default values.